refactor: decompose MainWindow.xaml.cs into feature-based partials
- Extracted Camera state and event handlers to [MainWindow.Camera.cs](file:///d:/Development/rustplus-desktop/RustPlusDesktop/Views/MainWindow/Camera/MainWindow.Camera.cs)
- Extracted Device and Storage management logic to [MainWindow.Devices.cs](file:///d:/Development/rustplus-desktop/RustPlusDesktop/Views/MainWindow/Devices/MainWindow.Devices.cs)
- Extracted Overlay drawing tools and team overlay logic to [MainWindow.Overlay.cs](file:///d:/Development/rustplus-desktop/RustPlusDesktop/Views/MainWindow/Overlay/MainWindow.Overlay.cs)
- Extracted Map marker lifecycle and polling to [MainWindow.Map.Markers.cs](file:///d:/Development/rustplus-desktop/RustPlusDesktop/Views/MainWindow/Map/MainWindow.Map.Markers.cs)
- Resolved build errors by fixing missing namespace imports and static method references
- Significantly reduced the complexity and size of the monolithic [MainWindow.xaml.cs](file:///d:/Development/rustplus-desktop/RustPlusDesktop/MainWindow.xaml.cs)
